Use langs in action, add `languages` input to manual trigger
Now uses `npm run test` as the action step. The language building
step should not do anything if there are no languages given in the
SECRET or in the manual workflow trigger. Input 'default' should
override languages in the repo SECRET.

Not sure `||` will work in setting env vars. There are docs about
that operator but it's unclear how exactly they function.